chore: rename vote to 24/05/2022
