+++ scripts/vote_2022_05_24.py
