+++ tests/test_2022_05_24_burn_shares.py
@@ -1,10 +1,10 @@
 """
-Tests for lido burnShares method for voting 17/05/2022
+Tests for lido burnShares method for voting 24/05/2022
 """
 import eth_abi
 import pytest
 from brownie import reverts, ZERO_ADDRESS, web3
-from scripts.vote_2022_05_17 import start_vote
+from scripts.vote_2022_05_24 import start_vote
 
 
 @pytest.fixture(scope="module")
