+++ tests/test_2022_05_24_pause_role_tests.py
@@ -1,10 +1,10 @@
 """
-Tests for lido resume/pause roles 17/05/2022
+Tests for lido resume/pause roles 24/05/2022
 """
 import pytest
 from functools import partial
 from brownie import reverts, web3
-from scripts.vote_2022_05_17 import start_vote
+from scripts.vote_2022_05_24 import start_vote
 
 
 @pytest.fixture(scope="module")
